backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pymongo==4.3.3
python-dotenv==1.0.0
motor==3.1.2
colorlog==6.7.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
cryptography==40.0.2